fix: refactor account timelines
