+++ src/app/user/[acct]/page.tsx
@@ -1,10 +1,11 @@
-import AccountTimeline from "@/components/AccountTimelnie";
+import AccountTimeline from "@/components/AccountTimeline";
 import UserHeader from "@/components/UserHeader";
 import { Account } from "@/models/account";
 import { DefaultInstance, HTTPError } from "@/models/api";
 import { stripPrefix } from "@/utils";
 import { Metadata } from "next";
 import { notFound } from "next/navigation";
+import { Suspense } from "react";
 
 async function getData(acct: string): Promise<{ readonly account: Account; readonly fullAcct: string }> {
   try {
@@ -37,7 +38,9 @@ export default async function UserPage({
   return (
     <>
       <UserHeader account={account} fullAcct={fullAcct} />
-      <AccountTimeline accountId={account.id} key={account.id} />
+      <Suspense fallback={<p>Loading...</p>}>
+        <AccountTimeline accountId={account.id} />
+      </Suspense>
     </>
   );
 }
