--- src/components/AccountTimelnie/styles.module.scss
@@ -1,5 +1,0 @@
-.staticTimeline {
-  & > li {
-    border-bottom: 1px solid var(--theme-status-border);
-  }
-}
