+++ src/rpc/app.ts
@@ -1,9 +1,9 @@
 import { DefaultInstance, FormDataRequestBody, SearchParamsRequestBody } from "@/models/api";
-import { buildAuthorizeUrl, revokeToken } from "@/models/api/mastodon/apps";
+import { revokeToken } from "@/models/api/mastodon/apps";
 import { getStatusesForAccount } from "@/models/api/mastodon/status";
 import { HomeTimelineRequestParamsZ, homeTimeline } from "@/models/api/mastodon/timeline";
-import { AppData, CreateAppRequest } from "@/models/app";
-import { getAuthorizationToken, setAuthorizationToken } from "@/models/auth";
+import { CreateAppRequest } from "@/models/app";
+import { getAuthorizationToken, getLoginUrl, setAuthorizationToken } from "@/models/auth";
 import { TRPCError, inferAsyncReturnType, initTRPC } from "@trpc/server";
 import { CreateNextContextOptions } from "@trpc/server/adapters/next";
 import pino from "pino";
@@ -50,15 +50,7 @@ export const appRpcRouter = t.router({
     );
     ctx.clearAuthorizedToken();
   }),
-  loginUrl: stdProcedure.query(async () => {
-    const app = await DefaultInstance.queryAppInfo(instance => instance.send(CreateAppRequest));
-    return buildAuthorizeUrl(DefaultInstance, {
-      response_type: "code",
-      client_id: app.client_id,
-      redirect_uri: AppData.redirect_uris,
-      scope: AppData.scopes,
-    });
-  }),
+  loginUrl: stdProcedure.query(getLoginUrl),
   account: t.router({
     statuses: stdProcedure
       .input(z.object({ accountId: z.string(), max_id: z.string().optional(), limit: z.number().optional() }))
