--- src/utils/cookie.ts
@@ -1,7 +1,0 @@
-import { cookies } from "next/headers";
-
-export namespace cookie {
-  export function token(): string | undefined {
-    return cookies().get("_lla")?.value;
-  }
-}
